requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
pydantic
requests
python-multipart
curl-cffi==0.2.1